last working file before Shayne stuff

+++ Student.java
@@ -1,9 +1,11 @@
+import java.text.DecimalFormat;
+
 public class Student {
     private String emailName;
     private String actualClass;
-    private float spamProbability;
+    private double spamProbability;
     
-    public Student(String emailName, String actualClass, float spamProbability) {
+    public Student(String emailName, String actualClass, double spamProbability) {
         this.emailName = emailName;
         this.actualClass = actualClass;
         this.spamProbability = spamProbability;
@@ -11,9 +13,9 @@ public class Student {
     
     public String getEmailName() { return this.emailName; }
     public String getActualClass() { return this.actualClass; }
-    public float getSpamProbability() { return this.spamProbability; }
+    public double getSpamProbability() { return this.spamProbability; }
     
     public void setEmailName(String emailName) { this.emailName = emailName; }
     public void setActualClass(String actualClass) { this.actualClass = actualClass; }
-    public void setSpamProbability(float spamProbability) { this.spamProbability = spamProbability; }
+    public void setSpamProbability(double spamProbability) { this.spamProbability = spamProbability; }
 }
